Middleware and admin authorization

+++ routes/api/admins.js
@@ -9,6 +9,7 @@ const Partner = require("../../models/Partner");
 const Project = require("../../models/Project");
 const Certificate = require("../../models/Certificate");
 const validator = require("../../validations/adminValidations");
+const passport = require('passport')
 //Create admin profile
 router.post("/register", async (req, res) => {
   try {
@@ -44,10 +45,10 @@ router.post("/register", async (req, res) => {
     });
   }
 });
-//View admin profile by id
-router.get("/:id", async (req, res) => {
+//View my profile by id
+router.get("/profile", passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
-    const admin = await Admin.findById(req.params.id);
+    const admin = await Admin.findById(req.id);
     if (!admin)
       return res.status(404).send({
         error: "This profile does not exist"
@@ -62,9 +63,9 @@ router.get("/:id", async (req, res) => {
   }
 });
 //Update admin profile by id
-router.put("/:id", async (req, res) => {
+router.put("/updateProfile",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
-    const admin = await Admin.findById(req.params.id);
+    const admin = await Admin.findById(req.id);
     if (!admin)
       return res.status(404).send({
         error: "This profile does not exist"
@@ -87,7 +88,7 @@ router.put("/:id", async (req, res) => {
         email: req.body.email
       }, function (err) {
         if (!err) {
-          Admin.findByIdAndUpdate(req.params.id, req.body, {
+          Admin.findByIdAndUpdate(req.id, req.body, {
             new: true
           }, function (
             err,
@@ -108,7 +109,7 @@ router.put("/:id", async (req, res) => {
           });
       });
     } else {
-      await Admin.findByIdAndUpdate(req.params.id, req.body, {
+      await Admin.findByIdAndUpdate(req.id, req.body, {
         new: true
       }, function (
         err,
@@ -131,9 +132,13 @@ router.put("/:id", async (req, res) => {
   }
 });
 //Delete admin profile by id
-router.delete("/:id", async (req, res) => {
+router.delete("/delete",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
-    const deletedAdmin = await Admin.findByIdAndDelete(req.params.id);
+    const deletedAdmin = await Admin.findByIdAndDelete(req.id);
+    if (!deletedAdmin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     res.json({
       msg: "Your account has been deleted successfully",
       data: deletedAdmin
@@ -145,56 +150,125 @@ router.delete("/:id", async (req, res) => {
   }
 });
 //View all admins profiles
-router.get("/get/admins", async (req, res) => {
+router.get("/get/admins",passport.authenticate('jwt', {session: false}), async (req, res) => {
+  try{
+  const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const admins = await Admin.find();
   res.json({
     data: admins
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //View all candidates profiles
-router.get("/get/candidates", async (req, res) => {
+router.get("/get/candidates",passport.authenticate('jwt', {session: false}), async (req, res) => {
+ try{
+  const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const candidates = await Candidate.find();
   res.json({
     data: candidates
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //View all consultancies profiles
-router.get("/get/consultancies", async (req, res) => {
+router.get("/get/consultancies",passport.authenticate('jwt', {session: false}), async (req, res) => {
+  try{
+  const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const consultancies = await Consultancy.find();
   res.json({
     data: consultancies
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //View all partners profiles
-router.get("/get/partners", async (req, res) => {
+router.get("/get/partners",passport.authenticate('jwt', {session: false}), async (req, res) => {
+ try{
+  const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const partners = await Partner.find();
   res.json({
     data: partners
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //View all certificates
-router.get("/get/certificates", async (req, res) => {
+router.get("/get/certificates",passport.authenticate('jwt', {session: false}), async (req, res) => {
+  try{
+  const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const certificates = await Certificate.find();
   res.json({
     data: certificates
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //View all projects
-router.get("/get/projects", async (req, res) => {
+router.get("/get/projects",passport.authenticate('jwt', {session: false}), async (req, res) => {
+ try{
+  const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const projects = await Project.find();
   res.json({
     data: projects
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //View all emails
-router.get("/get/emails", async (req, res) => {
+router.get("/get/emails",passport.authenticate('jwt', {session: false}), async (req, res) => {
+ try{
+  const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const emails = await Email.find();
   res.json({
     data: emails
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //search admins by name not exact value (search engine)
-router.get('/searchAdmins/:name', async (req, res) => {
+router.get('/searchAdmins/:name',passport.authenticate('jwt', {session: false}), async (req, res) => {
+try{
+  const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const admins = await Admin.find({
     name: {
       $regex: new RegExp(req.params.name)
@@ -202,10 +276,19 @@ router.get('/searchAdmins/:name', async (req, res) => {
   });
   res.json({
     data: admins
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //search candidates by name not exact value (search engine)
-router.get('/searchCandidates/:name', async (req, res) => {
+router.get('/searchCandidates/:name',passport.authenticate('jwt', {session: false}), async (req, res) => {
+  try{
+  const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const candidates = await Candidate.find({
     name: {
       $regex: new RegExp(req.params.name)
@@ -213,10 +296,19 @@ router.get('/searchCandidates/:name', async (req, res) => {
   });
   res.json({
     data: candidates
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //search consultancies by name not exact value (search engine)
-router.get('/searchConsultancies/:name', async (req, res) => {
+router.get('/searchConsultancies/:name',passport.authenticate('jwt', {session: false}), async (req, res) => {
+ try{
+  const admin = await Admin.findById(req.id);
+  if (!admin)
+    return res.status(404).send({
+      error: "This profile does not exist"
+    });
   const consultancies = await Consultancy.find({
     name: {
       $regex: new RegExp(req.params.name)
@@ -224,10 +316,19 @@ router.get('/searchConsultancies/:name', async (req, res) => {
   });
   res.json({
     data: consultancies
-  });
+  });}
+catch(err){
+  return res.json(err.message)
+}
 });
 //search partners by name not exact value (search engine)
-router.get('/searchPartners/:name', async (req, res) => {
+router.get('/searchPartners/:name',passport.authenticate('jwt', {session: false}), async (req, res) => {
+  try{
+  const admin = await Admin.findById(req.id);
+  if (!admin)
+    return res.status(404).send({
+      error: "This profile does not exist"
+    });
   const partners = await Partner.find({
     name: {
       $regex: new RegExp(req.params.name)
@@ -235,10 +336,19 @@ router.get('/searchPartners/:name', async (req, res) => {
   });
   res.json({
     data: partners
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //search certificates by name not exact value (search engine)
-router.get('/searchCertificates/:name', async (req, res) => {
+router.get('/searchCertificates/:name',passport.authenticate('jwt', {session: false}), async (req, res) => {
+  try{
+  const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const certificates = await Certificate.find({
     name: {
       $regex: new RegExp(req.params.name)
@@ -246,10 +356,19 @@ router.get('/searchCertificates/:name', async (req, res) => {
   });
   res.json({
     data: certificates
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //search projects by name not exact value (search engine)
-router.get('/searchProjects/:name', async (req, res) => {
+router.get('/searchProjects/:name',passport.authenticate('jwt', {session: false}),async (req, res) => {
+try{
+  const admin = await Admin.findById(req.id);
+  if (!admin)
+    return res.status(404).send({
+      error: "This profile does not exist"
+    });
   const projects = await Project.find({
     name: {
       $regex: new RegExp(req.params.name)
@@ -257,10 +376,19 @@ router.get('/searchProjects/:name', async (req, res) => {
   });
   res.json({
     data: projects
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //search emails by name not exact value (search engine)
-router.get('/searchEmails/:name', async (req, res) => {
+router.get('/searchEmails/:name',passport.authenticate('jwt', {session: false}), async (req, res) => {
+  try{
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
   const emails = await Email.find({
     email: {
       $regex: new RegExp(req.params.name)
@@ -268,12 +396,20 @@ router.get('/searchEmails/:name', async (req, res) => {
   });
   res.json({
     data: emails
-  });
+  });}
+  catch(err){
+    return res.json(err.message)
+  }
 });
 //Create a new conversation by stating receiver email
-router.post("/conversation/:id", async (req, res) => {
+router.post("/conversations/start", passport.authenticate('jwt', {session: false}),async (req, res) => {
   try {
-    const senderID = req.params.id;
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
+    const senderID = req.id;
     const receiverEmail = req.body.email;
     if (!receiverEmail)
       return res.status(404).send({
@@ -374,9 +510,9 @@ router.post("/conversation/:id", async (req, res) => {
   }
 });
 //Get all my existing conversations
-router.get("/conversation/:id", async (req, res) => {
+router.get("/conversations/get", passport.authenticate('jwt', {session: false}),async (req, res) => {
   try {
-    const senderAdmin = await Admin.findById(req.params.id);
+    const senderAdmin = await Admin.findById(req.id);
     if (!senderAdmin)
       return res.status(404).send({
         error: "This profile does not exist"
@@ -391,9 +527,9 @@ router.get("/conversation/:id", async (req, res) => {
   }
 });
 //Get an existing conversation by stating receiver email
-router.get("/conversation/:id/:email", async (req, res) => {
+router.get("/conversations/get/:email",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
-    const senderAdmin = await Admin.findById(req.params.id);
+    const senderAdmin = await Admin.findById(req.id);
     if (!senderAdmin)
       return res.status(404).send({
         error: "This profile does not exist"
@@ -410,9 +546,14 @@ router.get("/conversation/:id/:email", async (req, res) => {
   }
 });
 //Delete an existing conversation by stating receiver email
-router.delete("/conversation/:id", async (req, res) => {
+router.delete("/conversations/delete",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
-    const senderID = req.params.id;
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
+    const senderID = req.id;
     const receiverEmail = req.body.email;
     if (!receiverEmail)
       return res.status(404).send({
@@ -512,9 +653,14 @@ router.delete("/conversation/:id", async (req, res) => {
   }
 });
 //send an email inside an existing conversation by stating receiver email and email content and email type
-router.post("/conversation/email/:id", async (req, res) => {
+router.post("/conversations/send",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
-    const senderID = req.params.id;
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
+    const senderID = req.id;
     const receiverEmail = req.body.email;
     const emailContent = req.body.content;
     const emailType = req.body.type;
@@ -618,8 +764,13 @@ router.post("/conversation/email/:id", async (req, res) => {
   }
 });
 //View an existing project by its id
-router.get("/project/:id", async (req, res) => {
+router.get("/project/:id", passport.authenticate('jwt', {session: false}),async (req, res) => {
   try {
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     const project = Project.findById(req.params.id);
     if (!project)
       return res.status(404).send({
@@ -635,8 +786,13 @@ router.get("/project/:id", async (req, res) => {
   }
 });
 //(Accept/Reject) a submitted project after negotiating and update all its attributes
-router.put("/project/:projectID", async (req, res) => {
+router.put("/project/:projectID",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     Project.findByIdAndUpdate(req.params.projectID, req.body, {
       new: true
     }, function (err, updatedProject) {
@@ -656,8 +812,13 @@ router.put("/project/:projectID", async (req, res) => {
   }
 });
 //Delete a submitted project by project id
-router.delete("/project/:projectID", async (req, res) => {
+router.delete("/project/:projectID",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     Project.findByIdAndDelete(req.params.projectID, function (err, deletedProject) {
       if (!err)
         Partner.update({
@@ -728,8 +889,13 @@ router.delete("/project/:projectID", async (req, res) => {
   }
 });
 //View all candidates applying for a by project by its id
-router.get("/projects/:projectID", async (req, res) => {
+router.get("/projects/:projectID",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     Project.findById(req.params.projectID, function (err, foundProject) {
       if (!err) {
         Candidate.find({
@@ -757,8 +923,13 @@ router.get("/projects/:projectID", async (req, res) => {
   }
 });
 //Approve a candidate by his id for a project he applied for by its id
-router.post("/project/:projectID/:candidateID", async (req, res) => {
+router.post("/project/:projectID/:candidateID",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     Project.findById(req.params.projectID, function (err, foundProject) {
       if (!err) {
         Candidate.update({
@@ -795,8 +966,13 @@ router.post("/project/:projectID/:candidateID", async (req, res) => {
   }
 });
 //create a new certificate
-router.post("/certificate", async (req, res) => {
+router.post("/certificate",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     const certificate = await Certificate.create(req.body);
     res.json({
       msg: certificate
@@ -808,8 +984,13 @@ router.post("/certificate", async (req, res) => {
   }
 });
 //View an existing certificate by it's id
-router.get("/certificate/:id", async (req, res) => {
+router.get("/certificate/:id", passport.authenticate('jwt', {session: false}),async (req, res) => {
   try {
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     const certificate = await Certificate.findById(req.params.id);
     if (!certificate)
       return res.status(404).send({
@@ -825,8 +1006,13 @@ router.get("/certificate/:id", async (req, res) => {
   }
 });
 //Update an existing certificate by it's id
-router.put("/certificate/:id", async (req, res) => {
+router.put("/certificate/:id", passport.authenticate('jwt', {session: false}),async (req, res) => {
   try {
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     Certificate.findByIdAndUpdate(req.params.id, req.body, {
       new: true
     }, function (err, updatedCertificate) {
@@ -846,8 +1032,13 @@ router.put("/certificate/:id", async (req, res) => {
   }
 });
 //Delete an existing certificate by it's id
-router.delete("/certificate/:id", async (req, res) => {
+router.delete("/certificate/:id", passport.authenticate('jwt', {session: false}),async (req, res) => {
   try {
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     const deletedCertificate = await Certificate.findByIdAndDelete(req.params.id);
     res.json({
       msg: "This certificate has been deleted successfully",
@@ -860,8 +1051,13 @@ router.delete("/certificate/:id", async (req, res) => {
   }
 });
 //Approve a candidate's evaluation by his id for a certficate he passed by it's id
-router.post("/certificate/:certificateID/:candidateID", async (req, res) => {
+router.post("/certificate/:certificateID/:candidateID",passport.authenticate('jwt', {session: false}), async (req, res) => {
   try {
+    const admin = await Admin.findById(req.id);
+    if (!admin)
+      return res.status(404).send({
+        error: "This profile does not exist"
+      });
     Certificate.findById(req.params.certificateID, function (err, foundCertificate) {
       if (!err) {
         Candidate.update({
